% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penguin_plot.R
\name{plot_species_distribution}
\alias{plot_species_distribution}
\title{Penguin Plot Histogram
A function to create a histogram for each species based on specified numeric variable}
\usage{
plot_species_distribution(data, variable_name)
}
\arguments{
\item{data}{A data frame containing information about different species, with at least one numeric variable and a variable named species to group by}

\item{variable}{A numeric variable to plot}
}
\value{
A histogram of the containing the summary variables (count, arithmetic mean, standard deviation, minimum and maximum value) for the chosen variable
}
\description{
Penguin Plot Histogram
A function to create a histogram for each species based on specified numeric variable
}
\examples{
Example usage:
Plot the distribution of 'body_mass_g' for each penguin species
}
