% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_species.R
\name{summarize_species}
\alias{summarize_species}
\title{Summarisation of species}
\usage{
summarize_species(data, variable)
}
\arguments{
\item{data}{A data frame containing information about different species, with at least one numeric variable and a variable named species to group by}
}
\value{
A dataframe containing the summary variables (count, arithmetic mean, standard deviation, minimum and maximum value) for the chosen variable
}
\description{
Summarise the count, arithmetic mean, standard deviation, and minimum and maximum values for a specified numeric variable in a dataframe dataset of species
}
\examples{
Example usage:
summarize_species(data = palmerpenguins::penguins,
               variable = flipper_length_mm)
}
